move basic_account to types module
